Update README.md
We are interested in opening an online shopping clothes store so we are going to collect information on how people are spending on clothes online. The information are collected are from surveys containing 17 questions. After exploring the data and working on the hypotheses, we will understand our customer to develop a marketing strategy for our targeted customers.